Removing Hydralit and use the news multi-page capability of streamlit
